Switch to 32bit Floats

+++ data/src/position.rs
@@ -18,16 +18,16 @@ impl Position {
         self.score
     }
 
-    pub fn result(&self) -> f64 {
-        f64::from(self.result) / 2.
+    pub fn result(&self) -> f32 {
+        f32::from(self.result) / 2.
     }
 
     pub fn result_idx(&self) -> usize {
         usize::from(self.result)
     }
 
-    pub fn blended_result(&self, blend: f64) -> f64 {
-        blend * self.result() + (1. - blend) * sigmoid(f64::from(self.score), 0.009)
+    pub fn blended_result(&self, blend: f32) -> f32 {
+        blend * self.result() + (1. - blend) * sigmoid(f32::from(self.score), 0.009)
     }
 
     pub fn stm(&self) -> usize {
