+++ trainer/src/optimiser.rs
@@ -3,28 +3,28 @@ use crate::arch::{NNUEParams, HIDDEN, INPUT};
 pub trait Optimiser: Default {
     fn update_feature_weights(
         &mut self,
-        params: &mut [f64; INPUT * HIDDEN],
-        grads: &[f64; INPUT * HIDDEN],
-        adj: f64,
-        rate: f64,
+        params: &mut [f32; INPUT * HIDDEN],
+        grads: &[f32; INPUT * HIDDEN],
+        adj: f32,
+        rate: f32,
     );
     fn update_feature_bias(
         &mut self,
-        params: &mut [f64; HIDDEN],
-        grads: &[f64; HIDDEN],
-        adj: f64,
-        rate: f64,
+        params: &mut [f32; HIDDEN],
+        grads: &[f32; HIDDEN],
+        adj: f32,
+        rate: f32,
     );
     fn update_output_weights(
         &mut self,
-        params: &mut [f64; 2 * HIDDEN],
-        grads: &[f64; 2 * HIDDEN],
-        adj: f64,
-        rate: f64,
+        params: &mut [f32; 2 * HIDDEN],
+        grads: &[f32; 2 * HIDDEN],
+        adj: f32,
+        rate: f32,
     );
-    fn update_output_bias(&mut self, param: &mut f64, grad: f64, adj: f64, rate: f64);
+    fn update_output_bias(&mut self, param: &mut f32, grad: f32, adj: f32, rate: f32);
 
-    fn update_weights(&mut self, nnue: &mut NNUEParams, grads: &NNUEParams, adj: f64, rate: f64) {
+    fn update_weights(&mut self, nnue: &mut NNUEParams, grads: &NNUEParams, adj: f32, rate: f32) {
         self.update_feature_weights(&mut nnue.feature_weights, &grads.feature_weights, adj, rate);
         self.update_output_weights(&mut nnue.output_weights, &grads.output_weights, adj, rate);
         self.update_feature_bias(&mut nnue.feature_bias, &grads.feature_bias, adj, rate);
@@ -38,10 +38,10 @@ pub struct Adam {
 }
 
 impl Adam {
-    const B1: f64 = 0.9;
-    const B2: f64 = 0.999;
+    const B1: f32 = 0.9;
+    const B2: f32 = 0.999;
 
-    fn update(p: &mut f64, m: &mut f64, v: &mut f64, grad: f64, rate: f64) {
+    fn update(p: &mut f32, m: &mut f32, v: &mut f32, grad: f32, rate: f32) {
         *m = Self::B1 * *m + (1. - Self::B1) * grad;
         *v = Self::B2 * *v + (1. - Self::B2) * grad * grad;
         *p -= rate * *m / (v.sqrt() + 0.000_000_01);
@@ -60,10 +60,10 @@ impl Default for Adam {
 impl Optimiser for Adam {
     fn update_feature_weights(
         &mut self,
-        params: &mut [f64; INPUT * HIDDEN],
-        grads: &[f64; INPUT * HIDDEN],
-        adj: f64,
-        rate: f64,
+        params: &mut [f32; INPUT * HIDDEN],
+        grads: &[f32; INPUT * HIDDEN],
+        adj: f32,
+        rate: f32,
     ) {
         for (i, param) in params.iter_mut().enumerate() {
             let grad = adj * grads[i];
@@ -79,10 +79,10 @@ impl Optimiser for Adam {
 
     fn update_feature_bias(
         &mut self,
-        params: &mut [f64; HIDDEN],
-        grads: &[f64; HIDDEN],
-        adj: f64,
-        rate: f64,
+        params: &mut [f32; HIDDEN],
+        grads: &[f32; HIDDEN],
+        adj: f32,
+        rate: f32,
     ) {
         for (i, param) in params.iter_mut().enumerate() {
             let grad = adj * grads[i];
@@ -98,10 +98,10 @@ impl Optimiser for Adam {
 
     fn update_output_weights(
         &mut self,
-        params: &mut [f64; 2 * HIDDEN],
-        grads: &[f64; 2 * HIDDEN],
-        adj: f64,
-        rate: f64,
+        params: &mut [f32; 2 * HIDDEN],
+        grads: &[f32; 2 * HIDDEN],
+        adj: f32,
+        rate: f32,
     ) {
         for (i, param) in params.iter_mut().enumerate() {
             let grad = adj * grads[i];
@@ -115,7 +115,7 @@ impl Optimiser for Adam {
         }
     }
 
-    fn update_output_bias(&mut self, param: &mut f64, mut grad: f64, adj: f64, rate: f64) {
+    fn update_output_bias(&mut self, param: &mut f32, mut grad: f32, adj: f32, rate: f32) {
         grad *= adj;
         Self::update(
             param,
